Type useCookiesWithOptions
